pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "get2work"
version = "0.1.0"
description = "The Python library that makes you actually work"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "gitpython>=3.1.0",
    "watchdog>=3.0.0",
    "psutil>=5.9.0",
    "anthropic>=0.20.0",
    "requests>=2.31.0",
    "apscheduler>=3.10.0",
    "pygame>=2.5.0",
]

[project.scripts]
get2work = "get2work.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["get2work"]